+++ src/router.js
@@ -259,6 +259,23 @@ export const routes = [
       }
     ]
   },
+  // --------------------------- //
+  //
+  // Plans
+  //
+  // --------------------------- //
+  {
+    name: 'plans',
+    path: '/plans',
+    component: () =>
+      import(/* webpackChunkName: "plans" */ '@/pages/Plans.vue'),
+    beforeEnter: multiguard([authNavGuard, tenantNavGuard])
+  },
+  // --------------------------- //
+  //
+  // Login
+  //
+  // --------------------------- //
   {
     name: 'login',
     path: '/login',